Extract symbol declaration and definitions
